Load document async
We'll need a loader or something here when we do this for real. I'd also
love to have the loading work on a background queue. I think that would
break all kinds of stuff since right now it's pretty tightly integrated
with Text Kit. It should be possible to do the document controller stuff
in the background and have text controller make sure it goes back to
main if needed from the document controller delegate messages.